requirements.txt
appdirs==1.4.4
autoflake==1.4
black==20.8b1
click==7.1.2
flake8==3.8.4
gitdb==4.0.5
GitPython==3.1.11
importlib-metadata==3.1.0
invoke==1.4.1
isort==5.6.4
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
regex==2020.11.13
smmap==3.0.4
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
